add authentication frontend and firebase config
