+++ client/src/App.jsx
@@ -1,0 +1,18 @@
+import React, { Component } from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import SignIn from "./components/Authentication/SignIn.jsx";
+import SignUp from "./components/Authentication/SignUp.jsx";
+import HomePage from "./HomePage.jsx";
+
+function App() {
+  return (
+    <Router>
+      <Routes>
+        <Route exact path="/" element={<SignIn />} />
+        <Route exact path="/signup" element={<SignUp />} />
+        <Route exact path="/home" element={<HomePage />} />
+      </Routes>
+    </Router>
+  );
+}
+export default App;
