+++ package.json
@@ -36,6 +36,7 @@
     "date-fns": "^2.29.3",
     "dayjs": "^1.11.6",
     "express": "^4.18.1",
+    "firebase": "^9.15.0",
     "jquery": "^3.6.0",
     "moment": "^2.29.4",
     "mongoose": "^6.3.5",
@@ -62,6 +63,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "prettier": "^2.7.1",
     "pretty-quick": "^3.1.3",
+    "react-router-dom": "^6.6.1",
     "style-loader": "^3.3.1",
     "webpack": "^5.73.0",
     "webpack-cli": "^4.9.2",
